separate login/logout from Partner ops
move login/logout code to LoginClient to improve readability and maintainability.
