+++ src/main/java/com/salesforce/dataloader/client/PartnerClient.java
@@ -41,43 +41,34 @@ import com.salesforce.dataloader.controller.Controller;
 import com.salesforce.dataloader.dyna.ParentIdLookupFieldFormatter;
 import com.salesforce.dataloader.dyna.SforceDynaBean;
 import com.salesforce.dataloader.exception.ParameterLoadException;
-import com.salesforce.dataloader.exception.PasswordExpiredException;
 import com.salesforce.dataloader.exception.RelationshipFormatException;
 import com.salesforce.dataloader.mapping.LoadMapper;
 import com.salesforce.dataloader.model.TableRow;
 import com.salesforce.dataloader.util.AppUtil;
-import com.sforce.soap.partner.Connector;
 import com.sforce.soap.partner.DeleteResult;
 import com.sforce.soap.partner.DescribeGlobalResult;
 import com.sforce.soap.partner.DescribeGlobalSObjectResult;
 import com.sforce.soap.partner.DescribeSObjectResult;
 import com.sforce.soap.partner.Error;
 import com.sforce.soap.partner.Field;
-import com.sforce.soap.partner.GetUserInfoResult;
 import com.sforce.soap.partner.LimitInfo;
 import com.sforce.soap.partner.LimitInfoHeader_element;
-import com.sforce.soap.partner.LoginResult;
 import com.sforce.soap.partner.PartnerConnection;
 import com.sforce.soap.partner.QueryResult;
 import com.sforce.soap.partner.SaveResult;
 import com.sforce.soap.partner.UndeleteResult;
 import com.sforce.soap.partner.UpsertResult;
 import com.sforce.soap.partner.fault.ApiFault;
-import com.sforce.soap.partner.fault.ExceptionCode;
 import com.sforce.soap.partner.fault.LoginFault;
-import com.sforce.soap.partner.fault.UnexpectedErrorFault;
 import com.sforce.soap.partner.sobject.SObject;
 import com.sforce.ws.ConnectionException;
 import com.sforce.ws.ConnectorConfig;
-import com.sforce.ws.SessionRenewer;
 
 import org.apache.commons.beanutils.DynaBean;
 import org.apache.logging.log4j.Logger;
 import com.salesforce.dataloader.util.DLLogManager;
 
 import java.lang.reflect.InvocationTargetException;
-import java.net.MalformedURLException;
-import java.net.URL;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
@@ -88,8 +79,6 @@ import java.util.TreeMap;
 public class PartnerClient extends ClientBase<PartnerConnection> {
 
     private static Logger LOG = DLLogManager.getLogger(PartnerClient.class);
-
-    private ConnectorConfig connectorConfig = null;
 
     private static interface ClientOperation<RESULT, ARG> {
         String getName();
@@ -196,18 +185,6 @@ public class PartnerClient extends ClientBase<PartnerConnection> {
         }
     };
 
-    private final ClientOperation<LoginResult, PartnerConnection> LOGIN_OPERATION = new ClientOperation<LoginResult, PartnerConnection>() {
-        @Override
-        public String getName() {
-            return "login";
-        }
-
-        @Override
-        public LoginResult run(PartnerConnection client) throws ConnectionException {
-            ConnectorConfig cc = client.getConfig();
-            return client.login(cc.getUsername(), cc.getPassword());
-        }
-    };
 
     private final ClientOperation<DescribeGlobalResult, Object> DESCRIBE_GLOBAL_OPERATION = new ClientOperation<DescribeGlobalResult, Object>() {
         @Override
@@ -240,10 +217,6 @@ public class PartnerClient extends ClientBase<PartnerConnection> {
     private final Map<String, ReferenceEntitiesDescribeMap> parentDescribeCache = new TreeMap<>(String.CASE_INSENSITIVE_ORDER);
     public PartnerClient(Controller controller) {
         super(controller, LOG);
-    }
-
-    public boolean connect() throws ConnectionException {
-        return login();
     }
 
     private void setExportBatchSize() {
@@ -292,6 +265,15 @@ public class PartnerClient extends ClientBase<PartnerConnection> {
         );
 
         return true;
+    }
+    
+    public PartnerConnection getConnection() {
+    	PartnerConnection conn = super.getConnection();
+    	if (conn == null) {
+    		conn = controller.getLoginClient().getConnection();
+    		super.setConnection(conn);
+    	}
+        return conn;
     }
 
     public UpsertResult[] loadUpserts(List<DynaBean> dynaBeans) throws ConnectionException {
@@ -366,8 +348,8 @@ public class PartnerClient extends ClientBase<PartnerConnection> {
 
     protected <R, A> R runOperation(ClientOperation<R, A> op, A arg) throws ConnectionException {
         logger.debug(Messages.getFormattedString("Client.beginOperation", op.getName())); //$NON-NLS-1$
-        if (op != this.LOGIN_OPERATION && !isSessionValid()) {
-            connect();
+        if (!controller.getLoginClient().isSessionValid()) {
+        	controller.getLoginClient().connect();
         }
         ConnectionException connectionException = null;
         try {
@@ -576,185 +558,6 @@ public class PartnerClient extends ClientBase<PartnerConnection> {
         return null;
     }
 
-    boolean isSessionValid() {
-        if (appConfig.getBoolean(AppConfig.PROP_SFDC_INTERNAL) && appConfig.getBoolean(AppConfig.PROP_SFDC_INTERNAL_IS_SESSION_ID_LOGIN)) {
-            return true;
-        }
-        if (appConfig.getString(AppConfig.PROP_OAUTH_ACCESSTOKEN) != null && appConfig.getString(AppConfig.PROP_OAUTH_ACCESSTOKEN).trim().length() > 0) {
-            return true;
-        }
-        return isLoggedIn();
-    }
-
-    private boolean login() throws ConnectionException, ApiFault {
-        disconnect();
-        String origEndpoint = new String(appConfig.getAuthEndpointForCurrentEnv());
-        try {
-            dologin();
-            logger.debug("able to successfully invoke server APIs of version " + getAPIVersionForTheSession());
-        } catch (UnexpectedErrorFault fault) {
-            // attempt login with previous API version
-            if (fault.getExceptionCode() == ExceptionCode.UNSUPPORTED_API_VERSION
-                    && getAPIVersionForTheSession() != null 
-                    && !getAPIVersionForTheSession().equals(getPreviousAPIVersionInWSC())) {
-                logger.error("Failed to successfully invoke server APIs of version " + getAPIVersionForTheSession());
-                setAPIVersionForTheSession(getPreviousAPIVersionInWSC());
-                login(); 
-            } else {
-                logger.error("Failed to get user info using manually configured session id", fault);
-                throw fault;
-            }
-        } catch (ConnectionException e) {
-            String exceptionMessage = e.getMessage();
-            if (e instanceof LoginFault) {
-                LoginFault lf = (LoginFault)e;
-                exceptionMessage = lf.getExceptionMessage();
-            }
-            logger.warn(Messages.getMessage(this.getClass(), "failedUsernamePasswordAuth", 
-                                            origEndpoint, appConfig.getString(AppConfig.PROP_SELECTED_SERVER_ENVIRONMENT), exceptionMessage));
-            // attempt login with the default endpoint for the selected environment
-            if (!appConfig.isDefaultAuthEndpointForCurrentEnv(origEndpoint)) {
-                // retry with default endpoint URL only if user is attempting production login
-                appConfig.setAuthEndpointForCurrentEnv(appConfig.getDefaultAuthEndpointForCurrentEnv());
-                logger.info(Messages.getMessage(this.getClass(), "retryUsernamePasswordAuth", appConfig.getDefaultAuthEndpointForCurrentEnv(), appConfig.getString(AppConfig.PROP_SELECTED_SERVER_ENVIRONMENT)));
-                login();
-            } else {
-                logger.error("Failed to get user info using manually configured session id", e);
-                throw e;   
-            }
-        } finally {
-            // restore original value of Config.ENDPOINT property
-            appConfig.setAuthEndpointForCurrentEnv(origEndpoint);
-        }
-        return true; // exception thrown if there is an issue with login
-    }
-    
-    private boolean dologin() throws ConnectionException, ApiFault {
-        // Attempt the login giving the user feedback
-        logger.info(Messages.getString("Client.sforceLogin")); //$NON-NLS-1$
-        String oauthAccessToken = appConfig.getString(AppConfig.PROP_OAUTH_ACCESSTOKEN);
-        String serverUrl = appConfig.getAuthEndpointForCurrentEnv();
-        if (oauthAccessToken != null && oauthAccessToken.trim().length() > 0) {
-            serverUrl = appConfig.getString(AppConfig.PROP_OAUTH_INSTANCE_URL);
-        }
-        final ConnectorConfig cc = getLoginConnectorConfig(serverUrl);
-        boolean savedIsTraceMessage = cc.isTraceMessage();
-        cc.setTraceMessage(false);
-        // NOTE
-        // Create PartnerConnection instance only once to avoid issues with API restricted editions
-        // such as Professional Edition.
-        final PartnerConnection conn = Connector.newConnection(cc);
-        // identify the client as dataloader
-        conn.setCallOptions(ClientBase.getClientName(this.appConfig), null);
-
-        try {
-            if (oauthAccessToken != null && oauthAccessToken.trim().length() > 0) {
-                setConfiguredSessionId(conn, oauthAccessToken, null);
-            } else if (appConfig.getBoolean(AppConfig.PROP_SFDC_INTERNAL) && appConfig.getBoolean(AppConfig.PROP_SFDC_INTERNAL_IS_SESSION_ID_LOGIN)) {
-                setConfiguredSessionId(conn, appConfig.getString(AppConfig.PROP_SFDC_INTERNAL_SESSION_ID), null);
-            } else {
-                setSessionRenewer(conn);
-                loginInternal(conn);
-            }
-        } catch (Exception ex) {
-            throw ex;
-        } finally {
-            cc.setTraceMessage(savedIsTraceMessage);
-        }
-        return true;
-
-    }
-
-    private void setSessionRenewer(final PartnerConnection conn) {
-        conn.getConfig().setSessionRenewer(new SessionRenewer() {
-            @Override
-            public SessionRenewalHeader renewSession(ConnectorConfig connectorConfig) throws ConnectionException {
-                loginInternal(conn);
-                return null;
-            }
-        });
-    }
-
-    private void setConfiguredSessionId(PartnerConnection conn, String sessionId, GetUserInfoResult userInfo) throws ConnectionException {
-        logger.info("Using manually configured session id to bypass login");
-        conn.setSessionHeader(sessionId);
-        if (userInfo == null) {
-            userInfo = conn.getUserInfo(); // check to make sure we have a good connection
-        }
-        loginSuccess(conn, getAuthenticationHostDomainUrl(appConfig.getAuthEndpointForCurrentEnv()), userInfo);
-    }
-
-    private void loginInternal(final PartnerConnection conn) throws ConnectionException, PasswordExpiredException {
-        final ConnectorConfig cc = conn.getConfig();
-        cc.setRequestHeader(AppConfig.CLIENT_ID_HEADER_NAME, appConfig.getClientIDForCurrentEnv());
-        try {
-            logger.info(Messages.getMessage(getClass(), "sforceLoginDetail", cc.getAuthEndpoint(), cc.getUsername()));
-            LoginResult loginResult = runOperation(LOGIN_OPERATION, conn);
-            // if password has expired, throw an exception
-            if (loginResult.getPasswordExpired()) {
-                throw new PasswordExpiredException(Messages
-                        .getString("Client.errorExpiredPassword")); //$NON-NLS-1$
-            }
-            // update session id and service endpoint based on response
-            conn.setSessionHeader(loginResult.getSessionId());
-            String serverUrl = loginResult.getServerUrl();
-            String server = getAuthenticationHostDomainUrl(serverUrl);
-            if (appConfig.getBoolean(AppConfig.PROP_RESET_URL_ON_LOGIN)) {
-                cc.setServiceEndpoint(serverUrl);
-            }
-            loginSuccess(conn, server, loginResult.getUserInfo());
-        } catch (ConnectionException ex) {
-            String exceptionMessage = ex.getMessage();
-            if (ex instanceof LoginFault) {
-                LoginFault lf = (LoginFault)ex;
-                exceptionMessage = lf.getExceptionMessage();
-            }
-            logger.error(Messages.getMessage(getClass(), "loginError", cc.getAuthEndpoint(), exceptionMessage), ex);
-            throw ex;
-        }
-    }
-
-    private void loginSuccess(PartnerConnection conn, String serv, GetUserInfoResult userInfo) {
-        setConnection(conn);
-        setSession(conn.getSessionHeader().getSessionId(), serv, userInfo);
-    }
-
-    private String getServerStringFromUrl(URL url) {
-        return url.getProtocol() + "://" + url.getAuthority();
-    }
-
-    private String getAuthenticationHostDomainUrl(String serverUrl) {
-        if (appConfig.getBoolean(AppConfig.PROP_RESET_URL_ON_LOGIN)) {
-            try {
-                return getServerStringFromUrl(new URL(serverUrl));
-            } catch (MalformedURLException e) {
-                logger.fatal("Unexpected error", e);
-                throw new RuntimeException(e);
-            }
-        }
-        // Either RESET_URL_ON_LOGIN is false or input param serverURL 
-        // is invalid. Try returning configured Auth endpoint instead.
-        return appConfig.getAuthEndpointForCurrentEnv();
-    }
-
-    public boolean logout() {
-        try {
-            PartnerConnection pc = getConnection();
-            if (pc != null) pc.logout();
-            
-        } catch (ConnectionException e) {
-            // ignore
-        } finally {
-            disconnect();
-        }
-        return true;
-    }
-
-    public void disconnect() {
-        clearSession();
-        setConnection(null);
-    }
-
     /**
      * Set the map of references to object external id info for current entity
      *
@@ -886,17 +689,6 @@ public class PartnerClient extends ClientBase<PartnerConnection> {
         describeSObject(appConfig.getString(AppConfig.PROP_ENTITY));
     }
 
-    /**
-     * @return true if loggedIn
-     */
-    public boolean isLoggedIn() {
-        return getSession().isSessionValid();
-    }
-
-    public void validateSession() {
-        getSession().validate();
-    }
-
     @Override
     public ConnectorConfig getConnectorConfig() {
         ConnectorConfig cc = super.getConnectorConfig();
@@ -907,17 +699,6 @@ public class PartnerClient extends ClientBase<PartnerConnection> {
     public static String getServicePath() {
         return "/services/Soap/u/" + getAPIVersionForTheSession() + "/";
     }
-    
-    private synchronized ConnectorConfig getLoginConnectorConfig(String serverURL) {
-        if (serverURL == null || serverURL.isEmpty()) {
-            serverURL = appConfig.getAuthEndpointForCurrentEnv();
-        }
-        this.connectorConfig = getConnectorConfig();
-        this.connectorConfig.setAuthEndpoint(serverURL + getServicePath());
-        this.connectorConfig.setServiceEndpoint(serverURL + getServicePath());
-        return this.connectorConfig;
-    }
-
     /**
      * This function returns the describe call for an sforce entity
      *
