+++ src/main/resources/messages.properties
@@ -92,9 +92,9 @@ Client.invalidEndpoint=Invalid salesforce.com endpoint URL specified: {0}
 Client.targetError=Error in Target text, check the value.
 Client.bindingError=Error creating binding to soap service, error was: {0}
 Client.sforceLogin=Beginning Partner Salesforce login ....
-PartnerClient.sforceLoginDetail=Salesforce login to {0} as user {1}
-PartnerClient.failedUsernamePasswordAuth=Failed to authenticate using the authentication service URL {0} specified for {1} environment. Error: {2}
-PartnerClient.retryUsernamePasswordAuth=Retrying to authenticate using username/password server URL {0} by setting the property {1}
+LoginClient.sforceLoginDetail=Salesforce login to {0} as user {1}
+LoginClient.failedUsernamePasswordAuth=Failed to authenticate using the authentication service URL {0} specified for {1} environment. Error: {2}
+LoginClient.retryUsernamePasswordAuth=Retrying to authenticate using username/password server URL {0} by setting the property {1}
 AppUtil.sforceLoginProxyDetail=Calls to Salesforce service will use proxy host: {0} port: {1}
 AppUtil.sforceLoginProxyUser=Calls to Salesforce service will use proxy user: {0}
 AppUtil.sforceLoginProxyPassword=Calls to Salesforce service will use proxy password from settings
